Genui 2.0.2

+++ pages/main.tsx
@@ -25,7 +25,7 @@ export default function Main({
   let [promptValue, setPromptValue] = useState<string>("");
 
   let { sandpack, dispatch } = useSandpack();
-  let { updateFile, addFile, deleteFile } = sandpack;
+  let { updateFile, addFile, deleteFile, error } = sandpack;
 
   // Inspired by https://github.com/vercel/ai/blob/main/examples/next-openai/app/api/assistant/route.ts
   let waitForRun = async (run: Run) => {
@@ -45,37 +45,53 @@ export default function Main({
       throw new Error(run.status); // TODO: Show alert
     }
 
-    if (run.status === "requires_action") {
-      if (run.required_action?.type === "submit_tool_outputs") {
-        const tool_outputs =
-          run.required_action.submit_tool_outputs.tool_calls.map((toolCall) => {
-            const args = JSON.parse(toolCall.function.arguments);
-            if (toolCall.function.name === "updateFile") {
+    if (
+      run.status === "requires_action" &&
+      run.required_action?.type === "submit_tool_outputs"
+    ) {
+      const tool_outputs =
+        run.required_action.submit_tool_outputs.tool_calls.map((toolCall) => {
+          const args = JSON.parse(toolCall.function.arguments);
+          let output = {
+            success: true,
+            errors: [],
+          };
+          switch (toolCall.function.name) {
+            case "updateFile":
               updateFile(args.path, args.code);
-            } else if (toolCall.function.name === "addFile") {
+              break;
+            case "addFile":
               addFile(args.path, args.code);
-            } else if (toolCall.function.name === "addFile") {
+              break;
+            case "deleteFile":
               deleteFile(args.path);
-            } else {
+              break;
+            case "getErrors":
+              output["errors"] = error ? [error] : [];
+              break;
+            case "readFile":
+              output["result"] = { content: sandpack.files[args.path].code };
+              break;
+            case "listFiles":
+              output["result"] = { paths: Object.keys(sandpack.files) };
+              break;
+            default:
               throw new Error(
                 `Unknown tool call function: ${toolCall.function.name}`
               );
-            }
-            return {
-              tool_call_id: toolCall.id,
-              output: JSON.stringify({
-                success: true,
-                errors: [], // TODO: Handle errors
-              }),
-            };
-          });
+          }
 
-        run = await submitToolOutputs(run.thread_id, run.id, {
-          tool_outputs,
+          return {
+            tool_call_id: toolCall.id,
+            output: JSON.stringify(output),
+          };
         });
 
-        await waitForRun(run);
-      }
+      run = await submitToolOutputs(run.thread_id, run.id, {
+        tool_outputs,
+      });
+
+      await waitForRun(run);
     }
   };
 
@@ -122,7 +138,6 @@ export default function Main({
       await waitForRun(initialRun);
     }
 
-    dispatch({ type: "refresh" });
     setIsGenerating(false);
   };
